requirements.txt
Flask>=3.0.0
paddleocr>=2.7.0
paddlepaddle>=3.0.0
Pillow>=10.0.0
opencv-python>=4.8.0